% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{as.percent}
\alias{as.percent}
\title{Format a proportion as a percentage string}
\usage{
as.percent(x, digits = 1, ...)
}
\arguments{
\item{x}{a numeric proportion, or vector of proportions, where \code{1}
represents one hundred percent}

\item{digits}{number of decimal places to show}

\item{...}{further arguments passed to \code{\link{formatC}}}
}
\value{
A character vector of percentage strings.
}
\description{
Turn a proportion (such as \code{0.75}) into a human-friendly percentage
string (such as \code{"75\%"}).
}
\examples{
as.percent(0.75)              # "75.0\\%"
as.percent(c(0.1, 0.005))     # "10.0\\%" "0.5\\%"
as.percent(2 / 3, digits = 0) # "67\\%"

}
\author{
Ben Wiseman, \email{benjamin.h.wiseman@gmail.com}
}
