% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%regex<-\%}
\alias{\%regex<-\%}
\title{Assign to a vector only where a regular expression matches}
\usage{
x \%regex<-\% value
}
\arguments{
\item{x}{a character vector}

\item{value}{a length-2 character vector of the form \code{c(pattern, replacement)}}
}
\value{
Used for the side effect of reassigning \code{x} in the calling
  environment; returns the modified \code{x} invisibly.
}
\description{
This takes two arguments, just like \code{gsub}: a pattern and a replacement.
It overwrites the \emph{entire} element wherever the pattern matches. If you
want to substitute only the matched portion, use \code{\%regex=\%} instead; to
replace matches with nothing (\code{""}), use \code{\%-\%} or \code{\%-=\%}.
}
\examples{
x <- c("a1b", "b1", "c", "d0")

# overwrite any element containing a digit
x \%regex<-\% c("\\\\d+", "x")
print(x)
# "x" "b" "c" "x"

}
\author{
Ben Wiseman, \email{benjamin.h.wiseman@gmail.com}
}
