% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{choose_permute}
\alias{choose_permute}
\alias{\%C\%}
\alias{\%P\%}
\title{Choose and permute operators}
\usage{
n \%C\% k

n \%P\% k
}
\arguments{
\item{n}{whole number (the \code{n} in "n choose/permute k")}

\item{k}{whole number (the \code{k} in "n choose/permute k")}
}
\value{
A numeric value.
}
\description{
Shorthand infix operators for common combinatorics: \code{n \%C\% k} gives the
number of combinations ("n choose k"), and \code{n \%P\% k} gives the number of
permutations ("n permute k").
}
\examples{
5 \%C\% 3 # 10  (5 choose 3)
5 \%P\% 3 # 60  (5 permute 3)

}
\author{
Ben Wiseman, \email{benjamin.h.wiseman@gmail.com}
}
