% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthand.R
\name{chr}
\alias{chr}
\alias{int}
\alias{dbl}
\alias{num}
\alias{bool}
\alias{as.class}
\title{Cleaner type-conversion functions}
\usage{
chr(x, ...)

int(x, ...)

dbl(x, ...)

num(x, ...)

bool(x, ...)

as.class(x, class)
}
\arguments{
\item{x}{value to be converted}

\item{...}{further arguments passed to the underlying \code{as.*} function}

\item{class}{character; the name of the class to convert \code{x} to (used by
\code{as.class})}
}
\value{
The value of \code{x} coerced to the requested type.
}
\description{
Short aliases for the most common \code{as.*} conversions. There is nothing
magical here, but the shorter names can make data-wrangling code much easier
to read, especially for users coming from other languages. \code{as.class()}
additionally converts to a class chosen by name at run time.
}
\examples{
chr(42)    # "42" = as.character()
int(42.1)  # 42L  = as.integer()
dbl("42")  # 42   = as.double()
num("42")  # 42   = as.numeric()
bool(42)   # TRUE = as.logical()


# as.class() converts to an arbitrary class chosen by name:
as.class(255, "roman") # CCLV
}
\author{
Steven Nydick, \email{steven.nydick@kornferry.com}

Ben Wiseman, \email{benjamin.h.wiseman@gmail.com}
}
