% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{comparisons}
\alias{comparisons}
\alias{\%==\%}
\alias{\%===\%}
\alias{\%>=\%}
\alias{\%<=\%}
\alias{\%><\%}
\alias{\%>=<\%}
\title{Comparison operators with better missing-value handling}
\usage{
x \%==\% y

x \%===\% y

x \%>=\% y

x \%<=\% y

x \%><\% y

x \%>=<\% y
}
\arguments{
\item{x}{a vector}

\item{y}{a vector (for the "between" operators, a length-2 vector of the form
\code{c(lower, upper)})}
}
\value{
A logical vector.
}
\description{
A set of comparison operators that improve on base R by treating missing
values as comparable (so two \code{NA}s are considered equal) and by adding
convenient interval ("between") tests. The operators are:

\itemize{
  \item \code{\%==\%} - equality that treats \code{NA == NA} as \code{TRUE}.
  \item \code{\%===\%} - strict equality of both value \emph{and} class, for
    those familiar with 'JavaScript' \code{===}.
  \item \code{\%>=\%}, \code{\%<=\%} - greater/less than or equal to, with
    missing-value equality.
  \item \code{\%><\%}, \code{\%>=<\%} - between, with the ends excluded or
    included respectively.
}

For approximate (floating-point) comparisons, see
\code{\link{floating_point_comparisons}}.
}
\examples{
 ## Equality and ordering, with missing-value equality
 c(1, NA, 3, 4)  ==  c(1, NA, 4, 3)
 #  TRUE    NA  FALSE FALSE

 c(1, NA, 3, 4) \%==\% c(1, NA, 4, 3)
 #  TRUE  TRUE  FALSE FALSE

 c(1, NA, 3, 4) \%>=\% c(1, NA, 4, 3)
 #  TRUE  TRUE FALSE  TRUE

 c(1, NA, 3, 4) \%<=\% c(1, NA, 4, 3)
 #  TRUE  TRUE  TRUE FALSE

 ## Strict equality - a la 'JavaScript' ===
 # Only TRUE if the class AND value of x and y are the same
 x <- int(2)
 y <- 2
 x == y         # TRUE
 x \%===\% y      # FALSE
 x \%===\% int(y) # TRUE

 ## Between
 # ends excluded
 2 \%><\% c(1, 3)  # TRUE
 3 \%><\% c(1, 3)  # FALSE

 # ends included
 2 \%>=<\% c(1, 3) # TRUE
 3 \%>=<\% c(1, 3) # TRUE

}
\seealso{
Other comparisons: 
\code{\link{floating_point_comparisons}}
}
\author{
Ben Wiseman, \email{benjamin.h.wiseman@gmail.com}
}
\concept{comparisons}
