% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{floating_point_comparisons}
\alias{floating_point_comparisons}
\alias{\%~=\%}
\alias{\%>~\%}
\alias{\%<~\%}
\title{Floating-point comparison operators}
\usage{
x \%~=\% y

x \%>~\% y

x \%<~\% y
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}
}
\value{
A logical value.
}
\description{
An important set of operators missing from base R. Using \code{==} on two
non-integer numbers can give unexpected results (see examples), because of the
way floating-point numbers are represented. These operators instead test
equality up to a small tolerance, via \code{\link[base]{all.equal}}.

For a fuller explanation, see
\url{https://docs.oracle.com/cd/E19957-01/806-3568/ncg_goldberg.html}.
}
\examples{
 ## Floating-point test of equality
 # base R:
 (0.1 + 0.1 + 0.1) == 0.3   # FALSE
 # with roperators:
 (0.1 + 0.1 + 0.1) \%~=\% 0.3 # TRUE

 # Note how the base >= and <= behave here:
 (0.1 + 0.1 + 0.1) \%>=\% 0.3 # TRUE
 (0.1 + 0.1 + 0.1) \%<=\% 0.3 # FALSE

 # Use \%>~\% and \%<~\% for greater/less than OR approximately equal
 (0.1 + 0.1 + 0.1) \%>~\% 0.3 # TRUE
 (0.1 + 0.1 + 0.1) \%<~\% 0.3 # TRUE

}
\seealso{
Other comparisons: 
\code{\link{comparisons}}
}
\author{
Ben Wiseman, \email{benjamin.h.wiseman@gmail.com}
}
\concept{comparisons}
