% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{\%~\%}
\alias{\%~\%}
\title{Fuzzy (case- and whitespace-insensitive) string equality}
\usage{
x \%~\% y
}
\arguments{
\item{x}{a character vector}

\item{y}{a character vector}
}
\value{
A logical vector, with \code{NA} where either side is \code{NA}.
}
\description{
The string counterpart to the numeric operator \code{\%~=\%} (see
\code{\link{floating_point_comparisons}}). \code{x \%~\% y} is \code{TRUE}
when \code{x} and \code{y} are equal ignoring case, leading/trailing
whitespace, and runs of internal whitespace - handy for joining or matching
messy, hand-entered data.
}
\examples{
"Foo " \%~\% "foo"                   # TRUE
"a  b" \%~\% "a b"                    # TRUE
c("Yes", "NO") \%~\% c("yes", "no")  # TRUE TRUE

}
\author{
Ben Wiseman, \email{benjamin.h.wiseman@gmail.com}
}
