% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{\%else\%}
\alias{\%else\%}
\title{Inline fallback for expressions that might error}
\usage{
expr \%else\% alternative
}
\arguments{
\item{expr}{an expression to try}

\item{alternative}{value (or expression) to return if \code{expr} errors}
}
\value{
The value of \code{expr}, or of \code{alternative} if \code{expr}
  raised an error.
}
\description{
Evaluate the left-hand side; if it raises an error, return the right-hand
side instead. The right-hand side is only evaluated when needed (lazily), so
an expensive or side-effecting fallback is safe to use.
}
\examples{
sqrt(4)         \%else\% NA_real_     # 2
sqrt("a")       \%else\% NA_real_     # NA, instead of an error
(1:3)[[99]]     \%else\% "out of range"
stop("boom")    \%else\% "recovered"

}
\author{
Ben Wiseman, \email{benjamin.h.wiseman@gmail.com}
}
