% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{integrate}
\alias{integrate}
\alias{\%integrate\%}
\title{Inline integration operator}
\usage{
f \%integrate\% range
}
\arguments{
\item{f}{a function with a numeric return value}

\item{range}{a length-2 numeric vector, \code{c(lower, upper)}}
}
\value{
A single numeric value: the value of the integral.
}
\description{
An inline call to \code{\link[stats]{integrate}} that returns the value of
the integral directly, rather than the usual list.
}
\examples{
f <- function(x) x^2
f \%integrate\% c(0, 1) # 0.3333333

# compared with base R, which returns a list:
str(integrate(f, 0, 1))

}
\author{
Ben Wiseman, \email{benjamin.h.wiseman@gmail.com}
}
