% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip_checks.R
\name{ip}
\alias{ip}
\alias{get_ip}
\title{IP checks}
\usage{
get_ip()
}
\description{
Quick ways to get your IP address from within R
EXPERIMENTAL for windows and linux only
}
\examples{
# determine IPV4
get_ip()


}
