% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{logicals}
\alias{logicals}
\alias{\%ni\%}
\alias{\%xor\%}
\alias{\%aon\%}
\title{Logical operators}
\usage{
x \%ni\% y

x \%xor\% y

x \%aon\% y
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
A logical vector.
}
\description{
A few convenience logical operators: "not in" (\code{\%ni\%}), exclusive or
(\code{\%xor\%}), and all-or-nothing (\code{\%aon\%}, which is \code{TRUE} when
\code{x} and \code{y} are both \code{TRUE} or both \code{FALSE}).
}
\examples{
 #### Not in ####
 "z" \%ni\% c("a", "b", "c") # TRUE

 #### Exclusive or ####
 TRUE  \%xor\% TRUE  # FALSE
 FALSE \%xor\% FALSE # FALSE
 FALSE \%xor\% TRUE  # TRUE

 #### All-or-nothing ####
 TRUE  \%aon\% TRUE  # TRUE
 FALSE \%aon\% FALSE # TRUE
 FALSE \%aon\% TRUE  # FALSE

}
\author{
Ben Wiseman, \email{benjamin.h.wiseman@gmail.com}
}
