% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthand.R
\name{n_unique}
\alias{n_unique}
\title{Count the number of unique values}
\usage{
n_unique(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector}

\item{na.rm}{logical; if \code{TRUE}, \code{NA}s are ignored when counting
unique values}
}
\value{
An integer count of unique values.
}
\description{
Returns the number of unique elements in \code{x}, optionally ignoring
\code{NA}s.
}
\examples{
n_unique(c(1, 2, 2, 3, NA))               # 4
n_unique(c(1, 2, 2, 3, NA), na.rm = TRUE) # 3

}
\author{
Ben Wiseman, \email{benjamin.h.wiseman@gmail.com}
}
