% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%==\%}
\alias{\%==\%}
\alias{\%>=\%}
\alias{\%<=\%}
\title{Test of equality including equality of missing values}
\usage{
x \%==\% y

x \%>=\% y

x \%<=\% y
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\description{
Test of equality including equality of missing values
}
\examples{
 c(1, NA, 3, 4)  ==  c(1, NA, 4, 3)
 #  TRUE    NA  FALSE FALSE

 c(1, NA, 3, 4) \%==\% c(1, NA, 4, 3)
 #  TRUE  TRUE  FALSE FALSE

 c(1, NA, 3, 4) \%>=\% c(1, NA, 4, 3)
 #  TRUE  TRUE FALSE  TRUE

 c(1, NA, 3, 4) \%<=\% c(1, NA, 4, 3)
 #  TRUE  TRUE TRUE  FALSE
}
