% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%regex=\%}
\alias{\%regex=\%}
\title{Modify an object in place by regular expression}
\usage{
x \%regex=\% value
}
\arguments{
\item{x}{a character vector}

\item{value}{a length-2 character vector of the form \code{c(pattern, replacement)}}
}
\value{
Used for the side effect of reassigning \code{x} in the calling
  environment; returns the modified \code{x} invisibly.
}
\description{
This takes two arguments, just like \code{gsub}: a pattern and a replacement.
It overwrites only the matched portion of each element of \code{x}. If you
want to overwrite whole elements that match (rather than just the matched
portion), use \code{\%regex<-\%} instead.
}
\examples{
x <- c("a1b", "b1", "c", "d0")

# change any digit to "x"
x \%regex=\% c("\\\\d+", "x")
print(x)
# "axb" "b" "c" "dx"

}
\author{
Ben Wiseman, \email{benjamin.h.wiseman@gmail.com}
}
