% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthand.R
\name{read.tsv}
\alias{read.tsv}
\alias{read.psv}
\title{Read tab- or pipe-separated files}
\usage{
read.tsv(file, ...)

read.psv(file, ...)
}
\arguments{
\item{file}{path of the file to load}

\item{...}{further arguments passed to \code{\link[utils]{read.table}}}
}
\value{
A \code{data.frame}.
}
\description{
Convenience wrappers around \code{\link[utils]{read.table}} for tab-separated
(\code{read.tsv}) and pipe-separated (\code{read.psv}) files. Both default to
\code{header = TRUE}, like \code{read.csv}.
}
