% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthand.R
\name{seq_around}
\alias{seq_around}
\title{Sequence of evenly spaced points around an origin}
\usage{
seq_around(origin = 1, n = 1, spacing = 0.25)
}
\arguments{
\item{origin}{number to centre the sequence on}

\item{n}{number of points to create (a single whole number)}

\item{spacing}{distance between any two neighbouring points}
}
\value{
A numeric vector. Defaults to \code{1} when called with no arguments,
  to mirror the default behaviour of \code{\link[base]{seq}}.
}
\description{
Returns a vector of \code{n} points evenly spaced around \code{origin}, with
the given \code{spacing} between neighbours.
}
\examples{
seq_around(0, n = 5, spacing = 1) # -2 -1  0  1  2
seq_around(10, n = 3)             # 9.75 10.00 10.25

}
\author{
Ben Wiseman, \email{benjamin.h.wiseman@gmail.com}
}
