% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{string_arithmetic}
\alias{string_arithmetic}
\alias{\%+\%}
\alias{\%-\%}
\alias{\%s*\%}
\alias{\%s/\%}
\title{String arithmetic operators}
\usage{
x \%+\% y

x \%-\% y

x \%s*\% y

x \%s/\% y
}
\arguments{
\item{x}{a string (character vector)}

\item{y}{a string (character vector or regular expression)}
}
\value{
A character vector for \code{\%+\%}, \code{\%-\%}, and \code{\%s*\%}, and an
  integer count for \code{\%s/\%}.
}
\description{
Perform string concatenation and arithmetic in a similar way to other
languages. String addition (\code{\%+\%}) glues strings together, string
subtraction (\code{\%-\%}) removes a pattern, string multiplication
(\code{\%s*\%}) repeats a string, and string division (\code{\%s/\%}) counts how
many times a pattern occurs. String division has no equivalent in languages
like 'Python', yet is arguably more useful than string multiplication, and it
accepts regular expressions.
}
\examples{
("ab" \%+\% "c") == "abc" # TRUE
("abc" \%-\% "b") == "ac" # TRUE
("ac" \%s*\% 2) == "acac" # TRUE
("acac" \%s/\% "c") == 2  # TRUE
# String division with a regular expression:
"an apple a day keeps the malignant spirit of Steve Jobs at bay" \%s/\% "Steve Jobs|apple"

}
\author{
Ben Wiseman, \email{benjamin.h.wiseman@gmail.com}
}
