% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%regex<-\%}
\alias{\%regex<-\%}
\title{Assign to vector only where regular expression is matched}
\usage{
x \%regex<-\% value
}
\arguments{
\item{x}{a character vector}

\item{value}{c(pattern, replacement)}
}
\description{
Assign to vector only where regular expression is matched
}
\examples{
# Overwrite elements that match regex:

 x <- c("a1b", "b1", "c", "d0")

 # overwrite any element containing a number

 x \%regex<-\% c("\\\\d+", "x")

 print(x)

 # "x" "b" "c" "x"
}
