% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{choose_permute}
\alias{choose_permute}
\alias{\%C\%}
\alias{\%P\%}
\title{Choose and permute}
\usage{
n \%C\% k

n \%P\% k
}
\arguments{
\item{n}{whole number (from n choose/permute k)}

\item{k}{whole number (from n choose/permute k)}
}
\description{
Shorthand for some common mathematical operators
}
\examples{
# Calculate 5 choose 3
print(5 \%C\% 3)
# Calculate 5 permute 3
print(5 \%P\% 3)

}
\author{
Ben Wiseman, \email{benjamin.wiseman@kornferry.com}
}
