% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_checks.R
\name{content_checks}
\alias{content_checks}
\alias{is.constant}
\alias{is.binary}
\title{Contents of Vector Checks}
\usage{
is.constant(x)

is.binary(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
a logical value
}
\description{
Misc/useful functions to easily determine what is contained in a vector.
}
\examples{
x <- rep(1, 10)
is.constant(x)
is.binary(x)

x <- c(rep_len(1:2, 10), NA)
is.constant(x)
is.binary(x)

x <- c(rep_len(1:3, 10), NA)
is.constant(x)
is.binary(x)
}
