% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{integrate}
\alias{integrate}
\alias{\%integrate\%}
\title{Inline integration}
\usage{
f \%integrate\% range
}
\arguments{
\item{f}{function (with numeric return)}

\item{range}{vector of two numbers c(low, high)}
}
\description{
inline call to integrate that returns integration value rather than list
}
\examples{
f <- function(x) x^2
print(f \%integrate\% c(0, 1))
# vs base
x <- integrate(f, 0, 1)
str(x)
}
\author{
Ben Wiseman, \email{benjamin.wiseman@kornferry.com}
}
