% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthand.R
\name{library.force}
\alias{library.force}
\alias{require.force}
\title{loads package if available, else tries to install it (from CRAN by default)}
\usage{
library.force(pkg, ...)

require.force(pkg, ...)
}
\arguments{
\item{pkg}{name of package to load/install}

\item{...}{other args used by install.packages}
}
\description{
loads package if available, else tries to install it (from CRAN by default)

loads package if available, else tries to install it (from CRAN by default)
}
