% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%ni\%}
\alias{\%ni\%}
\alias{\%xor\%}
\alias{\%aon\%}
\title{Logical operators}
\usage{
x \%ni\% y

x \%xor\% y

x \%aon\% y
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\description{
These are some convienience functions, such as a not-in, and xor operator.
}
\examples{

 #### Not in ####

 "z" \%ni\%  c("a", "b", "c")
 #  TRUE

 #### Exclusive or  ####

 TRUE \%xor\% TRUE
 # FALSE

 FALSE \%xor\% FALSE
 # FALSE

 FALSE \%xor\% TRUE
 # TRUE

 #### All-or-nothing ####

 TRUE \%aon\% TRUE
 # TRUE

 FALSE \%aon\% FALSE
 # TRUE

 FALSE \%aon\% TRUE
 # FALSE

}
