% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%regex=\%}
\alias{\%regex=\%}
\title{Modify or overwrite existing object by regular expression}
\usage{
x \%regex=\% value
}
\arguments{
\item{x}{a character vector}

\item{value}{c(pattern, replacement)}
}
\description{
Modify or overwrite existing object by regular expression
}
\examples{
# Apply a regular expression/substitution to x:

 x <- c("a1b", "b1", "c", "d0")

 # change any number to "x"

  x \%regex=\% c("\\\\d+", "x")

 print(x)

 # "axb" "b" "c" "dx"
}
