% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%na<-\%}
\alias{\%na<-\%}
\title{Assign value to a vector's missing values}
\usage{
x \%na<-\% value
}
\arguments{
\item{x}{a vector}

\item{value}{value to replace vector's missing values with}
}
\description{
\code{\%na<-\%} is a simple shortcut to assign a specific value to all
NA elements contained in x.
}
\examples{
 x <- c("a", NA, "c")

 x \%na<-\% "b"

 print(x)
 # "a" "b" "c"

 x <- c(1, NA, 3, NA)

 x \%na<-\% c(2,4)

 print(x)
 # 1 2 3 4

}
