% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorthand.R
\name{read.tsv}
\alias{read.tsv}
\alias{read.psv}
\title{Read TSV Files}
\usage{
read.tsv(file, ...)

read.psv(file, ...)
}
\arguments{
\item{file}{path of file you want to load}

\item{...}{other args used by read.table}
}
\description{
Like read.csv but for tsv and default header = TRUE.
}
