% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%+\%}
\alias{\%+\%}
\alias{\%-\%}
\alias{\%s*\%}
\alias{\%s/\%}
\title{String operators}
\usage{
x \%+\% y

x \%-\% y

x \%s*\% y

x \%s/\% y
}
\arguments{
\item{x}{a string}

\item{y}{a string}
}
\description{
Perform string concatenation and arithmetic is a similar way to other languages
}
\examples{
("ab" \%+\% "c") == "abc" # TRUE
("abc" \%-\% "b") == "ac" # TRUE
("ac" \%s*\% 2) == "acac" # TRUE
("acac" \%s/\% "c") == 2  # TRUE

}
